+++ R/pivot_longer.R
@@ -9,7 +9,9 @@
 #'  the start of each variable name.
 #' @param values_to A string specifying the name of the column to create from the
 #'  data stored in cell values.
-#' @inheritParams slice_tail.RPolarsDataFrame
+#' @inheritParams rlang::check_dots_empty0
+#'
+#' @inheritSection left_join.RPolarsDataFrame Unknown arguments
 #'
 #' @export
 #' @examplesIf require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)
@@ -29,11 +31,17 @@
 #'     names_prefix = "wk",
 #'     values_to = "rank",
 #'   )
-
 pivot_longer.RPolarsDataFrame <- function(data, cols, ..., names_to = "name",
-                                   names_prefix = NULL,
-                                   values_to = "value") {
-
+                                          names_prefix = NULL,
+                                          values_to = "value") {
+  check_dots_empty_ignore(
+    ...,
+    .unsupported = c(
+      "cols_vary", "names_sep", "names_pattern", "names_ptypes",
+      "names_transform", "names_repair", "values_drop_na", "values_ptypes",
+      "values_transform"
+    )
+  )
   data_names <- names(data)
   on <- tidyselect_named_arg(data, rlang::enquo(cols))
   index <- data_names[!data_names %in% on]
@@ -50,8 +58,8 @@ pivot_longer.RPolarsDataFrame <- function(data, cols, ..., names_to = "name",
     }
     out <- out$
       with_columns(
-        pl$col(names_to)$str$replace(paste0("^", names_prefix), "")
-      )
+      pl$col(names_to)$str$replace(paste0("^", names_prefix), "")
+    )
   }
 
   add_tidypolars_class(out)