+++ R/utils.R
@@ -36,24 +36,6 @@ parse_replacement <- function(x) {
   gsub("\\\\(\\d+)", "$\\1", x)
 }
 
-# mostly to make R CMD check happy about S3 consistency
-unused_args <- function(...) {
-  x <- get_dots(...)
-  unused <- c()
-  for (i in seq_along(x)) {
-    if (!is.null(eval(x[[i]], parent.frame(1L)))) {
-      unused <- c(unused, deparse(x[[i]]))
-    }
-  }
-  if (length(unused) == 0) {
-    return(invisible())
-  }
-  warn(
-    paste("Unused arguments:", toString(unused)),
-    call = caller_env(2)
-  )
-}
-
 get_grps <- function(.data, .by, env) {
   grps <- attributes(.data)$pl_grps
   inline_grps <- tidyselect_named_arg(.data, .by)
@@ -111,11 +93,81 @@ select_by_name_or_position <- function(expr, name, position, default, env) {
   }
 }
 
-check_unsupported_arg <- function(arg_name, ...) {
+# Check that dots are not used, but throw a custom error when some args are
+# supported by tidyverse but not tidypolars.
+check_dots_empty_ignore <- function(..., .unsupported = NULL) {
   dots <- enquos(...)
-  if (arg_name %in% names(dots)) {
+  rlang_action <- getOption("tidypolars_unknown_args", "warn")
+  unsupported_dots <- names(dots[.unsupported])
+  unsupported_dots <- unsupported_dots[!is.na(unsupported_dots)]
+
+  if (length(unsupported_dots) > 0) {
+    if (length(unsupported_dots) == 1) {
+      msg <- paste0("Argument `", unsupported_dots, "` is not supported by tidypolars.")
+    } else {
+      msg <- paste(
+        "Arguments",
+        toString(paste0("`", unsupported_dots, "`")),
+        "are not supported by tidypolars."
+      )
+    }
+    if (rlang_action == "warn") {
+      warn(msg, call = caller_env())
+    } else if (rlang_action == "error") {
+      abort(
+        c(
+          msg,
+          "i" = "Use `options(tidypolars_unknown_args = \"warn\")` to warn when this happens instead of throwing an error."
+        ),
+        call = caller_env()
+      )
+    }
+  }
+  for (i in .unsupported) {
+    dots[[i]] <- NULL
+  }
+
+  if (length(dots) > 0) {
     abort(
-      paste0("Argument `", arg_name, "` is not supported by tidypolars yet."),
+      c(
+        "`...` must be empty."
+        # "i" = paste("Unknown args:", dots))
+      ),
+      call = caller_env()
+    )
+  }
+}
+
+# mostly to make R CMD check happy about S3 consistency
+check_unsupported_arg <- function(...) {
+  dots <- list2(...)
+  if ("copy" %in% names(dots) && isFALSE(dots[["copy"]])) {
+    dots[["copy"]] <- NULL
+  }
+  unsupported_args <- names(Filter(Negate(is.null), dots))
+  rlang_action <- getOption("tidypolars_unknown_args", "warn")
+
+  if (length(unsupported_args) == 0) {
+    return(invisible())
+  }
+
+  if (length(unsupported_args) == 1) {
+    msg <- paste0("Argument `", unsupported_args, "` is not supported by tidypolars.")
+  } else {
+    msg <- paste(
+      "Arguments",
+      toString(paste0("`", unsupported_args, "`")),
+      "are not supported by tidypolars."
+    )
+  }
+  if (rlang_action == "warn") {
+    warn(msg, call = caller_env())
+  } else if (rlang_action == "error") {
+    abort(
+      c(
+        msg,
+        "i" = "Use `options(tidypolars_unknown_args = \"warn\")` to warn when this happens instead of throwing an error."
+      ),
       call = caller_env()
     )
   }