+++ tests/testthat/test-collect.R
@@ -29,3 +29,12 @@ test_that("can't collect non-LazyFrame object", {
     error = TRUE
   )
 })
+
+test_that("error on unknown args", {
+  pl_iris <- as_polars_lf(iris)
+
+  expect_snapshot(
+    collect(pl_iris, foo = TRUE),
+    error = TRUE
+  )
+})