+++ tests/testthat/test-pivot_longer-lazy.R
@@ -75,4 +75,23 @@ test_that("argument names_prefix works", {
   )
 })
 
+test_that("unsupported args throw warning", {
+  pl_billboard <- as_polars_lf(tidyr::billboard)
+
+  expect_warning(
+    pivot_longer(pl_billboard, cols_vary = "fastest", names_ptypes = TRUE)
+  )
+})
+
+test_that("dots must be empty", {
+  pl_billboard <- as_polars_lf(tidyr::billboard)
+
+  # Also gets the message from rlang because check_dots_used() is called before
+  # dispatching the generic
+  expect_snapshot_lazy(
+    pivot_longer(pl_billboard, foo = TRUE, cols_vary = "fastest"),
+    error = TRUE
+  )
+})
+
 Sys.setenv('TIDYPOLARS_TEST' = FALSE)